Renamed configuration interface to detach from plugins
Being configurable doesn't really have anything to do with the plugin
system, just that things in the plugin system should be configurable.
There are other things, like the factories, that want to be
configurable, but aren't plugins.
