+++ python/smqtk/utils/plugin.py
@@ -181,84 +181,6 @@ def make_config(plugin_getter):
     return d
 
 
-class ConfigurablePlugin (object):
-    """
-    Interface for plugin objects that should be configurable via a configuration
-    dictionary (think JSON).
-    """
-    __metaclass__ = abc.ABCMeta
-
-    @classmethod
-    def default_config(cls):
-        """
-        Generate and return a default configuration dictionary for this class.
-        This will be primarily used for generating what the configuration
-        dictionary would look like for this class without instantiating it.
-
-        By default, we observe what this class's constructor takes as arguments,
-        turning those argument names into configuration dictionary keys. If any
-        of those arguments have defaults, we will add those values into the
-        configuration dictionary appropriately. The dictionary returned should
-        only contain JSON compliant value types.
-
-        It is not be guaranteed that the configuration dictionary returned
-        from this method is valid for construction of an instance of this class.
-
-        :return: Default configuration dictionary for the class.
-        :rtype: dict
-
-        """
-        argspec = inspect.getargspec(cls.__init__)
-
-        # Ignores potential *args or **kwargs present
-        params = argspec.args[1:]  # skipping ``self`` arg
-        num_params = len(params)
-
-        if argspec.defaults:
-            num_defaults = len(argspec.defaults)
-            vals = ((None,) * (num_params - num_defaults) + argspec.defaults)
-        else:
-            vals = (None,) * num_params
-
-        return dict(zip(params, vals))
-
-    @classmethod
-    def from_config(cls, config_dict):
-        """
-        Instantiate a new instance of this class given the configuration
-        JSON-compliant dictionary encapsulating initialization arguments.
-
-        This method should not be called via super unless and instance of the
-        class is desired.
-
-        :param config_dict: JSON compliant dictionary encapsulating
-            a configuration.
-        :type config_dict: dict
-
-        """
-        # The simple case is that the class doesn't require any special
-        # parameters other than those that can be provided via the JSON
-        # specification, which we cover here. If an implementation needs
-        # something more special, they can override this function.
-        return cls(**config_dict)
-
-    @abc.abstractmethod
-    def get_config(self):
-        """
-        Return a JSON-compliant dictionary that could be passed to this class's
-        ``from_config`` method to produce an instance with identical
-        configuration.
-
-        In the common case, this involves naming the keys of the dictionary
-        based on the initialization argument names as if it were to be passed
-        to the constructor via dictionary expansion.
-
-        :return: JSON type compliant configuration dictionary.
-        :rtype: dict
-
-        """
-
-
 def to_plugin_config(cp_inst):
     """
     Helper method that transforms the configuration dictionary gotten from the
@@ -271,7 +193,7 @@ def to_plugin_config(cp_inst):
     TL;DR: This wraps the instance's ``get_config`` return in a certain way.
 
     :param cp_inst: Instance of a ConfigurablePlugin-subclass.
-    :type cp_inst: ConfigurablePlugin
+    :type cp_inst: Configurable
 
     :return: Plugin-format configuration dictionary.
     :rtype: dict
@@ -307,7 +229,7 @@ def from_plugin_config(config, plugin_getter, *args):
 
     :return: Instance of the configured class type as found in the given
         ``plugin_getter``.
-    :rtype: ConfigurablePlugin
+    :rtype: Configurable
 
     """
     t = config['type']
